[sosnode] Don't fail on sos-less bastion nodes used for node lists
If the master node is determined to not have sos installed, that is not
necessarily a fatal error for scenarios where the 'master' node is only
being used to enumerate node lists and is not actually part of the
cluster. This can happen when a user is using a bastion node to
enumerate and connect to the cluster environment, or if the local host
is being used to enumerate nodes via cluster client tooling.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>