[cirrus] Enable auto-cancellation of jobs and specify type for GCE
Modifies the cirrus configuration to enable auto-cancellation of jobs
for the master branch (it is already enabled for PR branches by
default). This will prevent batch merges from kicking off a test job for
each merge. While it may arise that two or more independent commits that
pass testing on their own branches combine to form an unexpected
failure, this possibility seems remote for sos.

This is a cost saving decision - currently for each commit to master we
are spinning up a minimum of 13 VMs, which can very quickly snowball if
we do multiple merges in a short amount of time (which is historically
how the project does merges).

Second, modify the `gce_instance` fields in the cirrus config to use
pre-defined machine types, as these match our current "custom"
specifications already, and are cheaper to run.

Resolves: #2541

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>