[docs] Clarify profile in manpage and how to enable further plugins
Updates the language in the manpage to clarify that in order to enable
additional plugins outside of a profile when a profile is specified,
that users must use `-o` instead of `-e`.

Remove the 'current profiles' list from the manpage as it is outdated
and it is better to refer users to the list reported by the local
installation.

Closes: #504
Resolves: #2543

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>